#17 Fix: Fix Data JPA Repository Mapping
- Data Jpa에서 잘못된 제네릭 매핑을 수정했습니다.

+++ linknamu/src/main/java/com/kakao/linknamu/category/CategoryJPARepository.java
@@ -2,6 +2,6 @@ package com.kakao.linknamu.category;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryJPARepository extends JpaRepository<Category, Integer> {
+public interface CategoryJPARepository extends JpaRepository<Category, Long> {
 
 }
