+++ linknamu/src/main/java/com/kakao/linknamu/tag/TagJPARepository.java
@@ -2,6 +2,6 @@ package com.kakao.linknamu.tag;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TagJPARepository extends JpaRepository<Tag, Integer> {
+public interface TagJPARepository extends JpaRepository<Tag, Long> {
 
 }
